♻️ ChatRoomService와 ChatMessageService의 findChatRoom 중복 로직을 관심사에 따라서 리팩토링 + 그에 따른 순환참조 문제 해결
