+++ src/main/java/com/sparta/taptoon/domain/chat/service/ChatMessageService.java
@@ -3,23 +3,21 @@ package com.sparta.taptoon.domain.chat.service;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.sparta.taptoon.domain.chat.dto.request.SendChatMessageRequest;
 import com.sparta.taptoon.domain.chat.dto.response.ChatCombinedMessageResponse;
-import com.sparta.taptoon.domain.chat.dto.response.ChatImageMessageResponse;
 import com.sparta.taptoon.domain.chat.dto.response.ChatMessageResponse;
 import com.sparta.taptoon.domain.chat.entity.ChatImageMessage;
 import com.sparta.taptoon.domain.chat.entity.ChatMessage;
 import com.sparta.taptoon.domain.chat.entity.ChatRoom;
 import com.sparta.taptoon.domain.chat.repository.ChatImageMessageRepository;
 import com.sparta.taptoon.domain.chat.repository.ChatMessageRepository;
-import com.sparta.taptoon.domain.chat.repository.ChatRoomRepository;
 import com.sparta.taptoon.domain.member.entity.Member;
 import com.sparta.taptoon.domain.member.repository.MemberRepository;
 import com.sparta.taptoon.global.error.enums.ErrorCode;
 import com.sparta.taptoon.global.error.exception.AccessDeniedException;
 import com.sparta.taptoon.global.error.exception.NotFoundException;
-import com.sparta.taptoon.global.handler.NotificationService;
 import com.sparta.taptoon.global.redis.RedisPublisher;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.bson.types.ObjectId;
 import org.springframework.data.redis.core.StringRedisTemplate;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -37,15 +35,13 @@ public class ChatMessageService {
     private static final String LAST_READ_MESSAGE_KEY_TEMPLATE = "chat:room:%s:user:%d";
 
     private final ChatMessageRepository chatMessageRepository;
-    private final ChatRoomRepository chatRoomRepository;
     private final MemberRepository memberRepository;
     private final RedisPublisher redisPublisher;
     private final ObjectMapper objectMapper;
     private final StringRedisTemplate redisTemplate;
-    private final SlackAlarmService slackAlarmService;
     private final ChatImageMessageRepository chatImageMessageRepository;
     private final ChatRoomService chatRoomService;
-    private final NotificationService notificationService;
+
 
     /**
      * 메시지 전송 & 저장, Redis와 Slack으로 알림을 발행합니다.
@@ -56,21 +52,15 @@ public class ChatMessageService {
      * @return 저장된 메시지의 응답 DTO
      */
     public ChatMessageResponse sendMessage(Long senderId, String chatRoomId, SendChatMessageRequest request) {
-        ChatRoom chatRoom = findChatRoom(chatRoomId);
+        ChatRoom chatRoom = chatRoomService.findChatRoom(chatRoomId);
         Member sender = findMember(senderId);
         validateChatRoomMembership(chatRoom, sender);
 
         ChatMessage chatMessage = saveChatMessage(request, chatRoom, sender);
         ChatMessageResponse response = ChatMessageResponse.from(chatMessage);
 
-        // 내가 보낸 메시지 읽음 처리
-        String key = String.format(LAST_READ_MESSAGE_KEY_TEMPLATE, chatRoomId, senderId);
-        redisTemplate.opsForValue().set(key, chatMessage.getId());
-        log.info("✅ 메시지 전송 및 읽음 처리 완료 - chatRoomId: {}, senderId: {}, messageId: {}",
-                chatRoomId, senderId, chatMessage.getId());
-
+        updateLastReadMessage(chatRoomId, senderId, chatMessage.getId());
         publishMessage(chatRoom.getId(), response, sender, request.message());
-        notificationService.notifyNewMessage(chatRoomId, senderId, request.message());
         return response;
     }
 
@@ -114,11 +104,11 @@ public class ChatMessageService {
      */
     @Transactional
     public List<ChatCombinedMessageResponse> getChatMessages(Long memberId, String chatRoomId) {
-        ChatRoom chatRoom = findChatRoom(chatRoomId);
+        ChatRoom chatRoom = chatRoomService.findChatRoom(chatRoomId);
         Member member = findMember(memberId);
         validateChatRoomMembership(chatRoom, member);
 
-        chatRoomService.updateUnreadMessages(chatRoom, memberId); // ChatRoomService로 이동
+        updateUnreadMessages(chatRoom, memberId); // 채팅방 진입 시 읽음 처리
 
         List<ChatMessage> textMessages = chatMessageRepository.findByChatRoomIdOrderByCreatedAtAsc(chatRoomId);
         List<ChatImageMessage> imageMessages = chatImageMessageRepository.findByChatRoomIdOrderByCreatedAtAsc(chatRoomId);
@@ -131,19 +121,61 @@ public class ChatMessageService {
                 .toList();
     }
 
-    // 채팅방 ID로 채팅방을 조회, 없으면 예외 발생
-    private ChatRoom findChatRoom(String chatRoomId) {
-        return chatRoomRepository.findById(chatRoomId)
-                .orElseThrow(() -> new NotFoundException(ErrorCode.CHAT_ROOM_NOT_FOUND));
+    // Redis에 마지막 읽은 메시지 저장
+    public void updateLastReadMessage(String chatRoomId, Long memberId, String messageId) {
+        String key = String.format(LAST_READ_MESSAGE_KEY_TEMPLATE, chatRoomId, memberId);
+        redisTemplate.opsForValue().set(key, messageId);
+        log.info("✅ 읽음 처리 - chatRoomId: {}, memberId: {}, messageId: {}", chatRoomId, memberId, messageId);
     }
 
-    // 사용자 ID로 사용자를 조회, 없으면 예외 발생
+    // 읽지 않은 메시지 수 계산
+    public int calculateUnreadCount(String chatRoomId, Long memberId) {
+        String key = String.format(LAST_READ_MESSAGE_KEY_TEMPLATE, chatRoomId, memberId);
+        String lastReadMessageIdStr = redisTemplate.opsForValue().get(key);
+        ObjectId lastReadMessageId = lastReadMessageIdStr != null ? new ObjectId(lastReadMessageIdStr) : null;
+
+        int unreadCount;
+        if (lastReadMessageId == null) {
+            unreadCount = chatMessageRepository.countByChatRoomId(chatRoomId);
+            log.info("No last read message for memberId: {}, counting all messages: {}", memberId, unreadCount);
+        } else {
+            unreadCount = chatMessageRepository.countUnreadMessagesExcludingSender(chatRoomId, lastReadMessageId, memberId);
+            log.info("Counting unread - chatRoomId: {}, memberId: {}, lastReadMessageId: {}, unreadCount: {}",
+                    chatRoomId, memberId, lastReadMessageId, unreadCount);
+        }
+        return unreadCount;
+    }
+
+    // 채팅방 조회 시 읽음 처리
+    @Transactional
+    public void updateUnreadMessages(ChatRoom chatRoom, Long memberId) {
+        String key = String.format(LAST_READ_MESSAGE_KEY_TEMPLATE, chatRoom.getId(), memberId);
+        String lastReadMessageIdStr = redisTemplate.opsForValue().get(key);
+        ObjectId lastReadMessageId = lastReadMessageIdStr != null ? new ObjectId(lastReadMessageIdStr) : null;
+
+        List<ChatMessage> unreadMessages = lastReadMessageId == null
+                ? chatMessageRepository.findByChatRoomIdOrderByCreatedAtAsc(chatRoom.getId())
+                : chatMessageRepository.findByChatRoomIdAndIdGreaterThan(chatRoom.getId(), lastReadMessageId);
+
+        if (!unreadMessages.isEmpty()) {
+            unreadMessages.forEach(message -> {
+                if (message.getUnreadCount() > 0 && !message.getSenderId().equals(memberId)) {
+                    message.decrementUnreadCount();
+                }
+            });
+            chatMessageRepository.saveAll(unreadMessages);
+
+            String latestMessageId = unreadMessages.get(unreadMessages.size() - 1).getId();
+            redisTemplate.opsForValue().set(key, latestMessageId);
+            log.info("✅ 읽음 처리 완료 - chatRoomId: {}, memberId: {}, latestMessageId: {}", chatRoom.getId(), memberId, latestMessageId);
+        }
+    }
+
     private Member findMember(Long memberId) {
         return memberRepository.findById(memberId)
                 .orElseThrow(() -> new NotFoundException(ErrorCode.CHAT_MEMBER_NOT_FOUND));
     }
 
-    // 사용자가 채팅방에 속해 있는지 검증, 속해 있지 않으면 예외 발생
     private void validateChatRoomMembership(ChatRoom chatRoom, Member member) {
         if (!chatRoom.getMemberIds().contains(member.getId())) {
             log.warn("Access denied for memberId: {} to chatRoomId: {}", member.getId(), chatRoom.getId());
@@ -152,7 +184,6 @@ public class ChatMessageService {
         log.info("Access granted for memberId: {} to chatRoomId: {}", member.getId(), chatRoom.getId());
     }
 
-    // 채팅 메시지를 생성하고 데이터베이스에 저장
     private ChatMessage saveChatMessage(SendChatMessageRequest request, ChatRoom chatRoom, Member sender) {
         ChatMessage chatMessage = ChatMessage.builder()
                 .chatRoomId(chatRoom.getId())
@@ -161,33 +192,14 @@ public class ChatMessageService {
                 .unreadCount(chatRoom.getMemberIds().size() - 1)
                 .build();
 
-        ChatMessage saved = chatMessageRepository.save(chatMessage);
-        return saved;
+        return chatMessageRepository.save(chatMessage);
     }
 
-    // Redis로 메시지를 발행하고 Slack으로 알림을 전송
     private void publishMessage(String chatRoomId, ChatMessageResponse response, Member sender, String message) {
         try {
             String jsonMessage = objectMapper.writeValueAsString(response);
             redisPublisher.publish(chatRoomId, jsonMessage);
             log.info("📤 Redis에 메시지 발행 완료: {}", response);
-
-            String slackMessage = String.format("📢 [채팅방 %s] %s: %s", chatRoomId, sender.getNickname(), message);
-            slackAlarmService.sendSlackMessage(slackMessage);
-        } catch (Exception e) {
-            log.error("❌ Redis 메시지 발행 중 오류 발생", e);
-        }
-    }
-
-    // Redis로 메시지를 발행하고 Slack으로 알림을 전송
-    private void publishImage(String chatRoomId, ChatImageMessageResponse response, Member sender, String imgUrl) {
-        try {
-            String jsonMessage = objectMapper.writeValueAsString(response);
-            redisPublisher.publish(chatRoomId, jsonMessage);
-            log.info("📤 Redis에 메시지 발행 완료: {}", response);
-
-            String slackMessage = String.format("📢 [채팅방 %d] %s: %s", chatRoomId, sender.getNickname(), imgUrl);
-            slackAlarmService.sendSlackMessage(slackMessage);
         } catch (Exception e) {
             log.error("❌ Redis 메시지 발행 중 오류 발생", e);
         }
@@ -197,28 +209,6 @@ public class ChatMessageService {
         return chatMessageRepository.findTopByChatRoomIdOrderByCreatedAtDesc(chatRoomId);
     }
 
-    // Redis에서 사용자의 마지막 읽은 메시지 ID를 조회
-    private String getLastReadMessageId(String chatRoomId, Long memberId) {
-        String key = String.format(LAST_READ_MESSAGE_KEY_TEMPLATE, chatRoomId, memberId);
-        String lastReadMessageIdStr = redisTemplate.opsForValue().get(key);
-        return lastReadMessageIdStr != null ? lastReadMessageIdStr : null;
-    }
-
-//    // 읽지 않은 메시지의 unreadCount를 감소시키고 Redis에 최신 읽은 메시지 ID를 업데이트
-//    private void updateUnreadMessages(ChatRoom chatRoom, Long memberId, String lastReadMessageId) {
-//        List<ChatMessage> unreadMessages = lastReadMessageId == null
-//                ? chatMessageRepository.findByChatRoomIdOrderByCreatedAtAsc(chatRoom.getId())
-//                : chatMessageRepository.findByChatRoomIdAndIdGreaterThan(chatRoom.getId(), lastReadMessageId);
-//        if (!unreadMessages.isEmpty()) {
-//            unreadMessages.forEach(ChatMessage::decrementUnreadCount);
-//            chatMessageRepository.saveAll(unreadMessages);
-//
-//            String latestMessageId = unreadMessages.get(unreadMessages.size() - 1).getId();
-//            String key = String.format(LAST_READ_MESSAGE_KEY_TEMPLATE, chatRoom.getId(), memberId);
-//            redisTemplate.opsForValue().set(key, latestMessageId);
-//            log.info("✅ 읽음 처리 완료 - chatRoomId: {}, memberId: {}, latestMessageId: {}", chatRoom.getId(), memberId, latestMessageId);
-//        }
-//    }
 
     // 채팅방의 모든 메시지를 시간순으로 조회하여 응답 DTO 리스트로 변환
     private List<ChatMessageResponse> fetchAllMessages(ChatRoom chatRoom) {
