+++ src/main/java/com/sparta/taptoon/global/handler/WebSocketHandler.java
@@ -3,7 +3,6 @@ package com.sparta.taptoon.global.handler;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.sparta.taptoon.domain.chat.dto.request.SendChatMessageRequest;
-import com.sparta.taptoon.domain.chat.entity.ChatMessage;
 import com.sparta.taptoon.domain.chat.service.ChatMessageService;
 import com.sparta.taptoon.global.error.exception.AccessDeniedException;
 import com.sparta.taptoon.global.error.exception.InvalidRequestException;
@@ -20,7 +19,6 @@ import org.springframework.web.socket.handler.TextWebSocketHandler;
 import java.io.IOException;
 import java.util.Collections;
 import java.util.Map;
-import java.util.Optional;
 import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
 
@@ -30,10 +28,8 @@ import java.util.concurrent.ConcurrentHashMap;
 public class WebSocketHandler extends TextWebSocketHandler {
 
     private static final String CHAT_ROOM_ID_PATH_INDEX = "/chat/";
-    private static final String LAST_READ_MESSAGE_KEY_TEMPLATE = "chat:room:%s:user:%d";
 
     private final ChatMessageService chatMessageService;
-    private final StringRedisTemplate redisTemplate;
     private final ObjectMapper objectMapper = new ObjectMapper();
 
     // 채팅방별 WebSocket 세션 관리
@@ -57,8 +53,6 @@ public class WebSocketHandler extends TextWebSocketHandler {
         chatRoomSessions.computeIfAbsent(chatRoomId, k -> ConcurrentHashMap.newKeySet()).add(session);
         log.info("✅ 채팅방 {} 에 세션 {} 추가 완료 (사용자: {})", chatRoomId, session.getId(), senderId);
 
-//        // 처음 접속시 읽음 처리
-//        updateLastReadMessageId(chatRoomId, senderId);
     }
 
     /**
@@ -115,9 +109,7 @@ public class WebSocketHandler extends TextWebSocketHandler {
         for (WebSocketSession session : sessions) {
             Long sessionSenderId = getSenderIdFromSession(session);
             if (sessionSenderId != null && sessionSenderId.equals(senderIdFromMessage)) {
-                String key = String.format(LAST_READ_MESSAGE_KEY_TEMPLATE, chatRoomId, sessionSenderId);
-                redisTemplate.opsForValue().set(key, messageId);
-                log.info("✅ 실시간 읽음 처리 - chatRoomId: {}, senderId: {}, messageId: {}", chatRoomId, sessionSenderId, messageId);
+                chatMessageService.updateLastReadMessage(chatRoomId, sessionSenderId, messageId);
             }
         }
 
@@ -182,17 +174,6 @@ public class WebSocketHandler extends TextWebSocketHandler {
         Long senderId = jsonNode.get("senderId").asLong();
         String message = jsonNode.get("message").asText();
         return new MessagePayload(senderId, message);
-    }
-
-    // 초기 접속 시 읽음 처리
-    private void updateLastReadMessageId(String chatRoomId, Long senderId) {
-        Optional<ChatMessage> latestMessage = chatMessageService.findLatestMessage(chatRoomId);
-        if (latestMessage.isPresent()) {
-            String key = String.format(LAST_READ_MESSAGE_KEY_TEMPLATE, chatRoomId, senderId);
-            redisTemplate.opsForValue().set(key, String.valueOf(latestMessage.get().getId()));
-            log.info("✅ 초기 읽음 처리 - chatRoomId: {}, senderId: {}, lastReadMessageId: {}",
-                    chatRoomId, senderId, latestMessage.get().getId());
-        }
     }
 
     // 메시지 페이로드 데이터 홀더
